+++ test/test_livingapi.py
@@ -3116,7 +3116,9 @@ def test_globals_dir(handler, config_data):
 			request=True
 			response=True
 			custom=True
+			form=True
 			geo=True
+			groups=True
 			dist=True
 			seq=True
 			flashes=True
@@ -3667,26 +3669,27 @@ def test_custom_attributes(handler, config_data):
 
 
 def test_user_change(handler, config_data):
-	vt = handler.make_viewtemplate(
-		la.DataSourceConfig(
-			identifier="persons",
-			app=config_data.apps.persons,
-		),
-		identifier="test_user_change",
-		source=f"""
-			<?print repr(globals.user.change("est123!", None, "nix@livinglogic.de"))?>
-			<?print repr(globals.user.change("Test123!", None, "nix@livinglogic.de"))?>
-			<?print globals.user.email?>
-			<?print repr(globals.user.change("Test123!", None, "xyz"))?>
-			<?code globals.user.change("Test123!", None, "walter@livinglogic.de")?>
-		""",
-	)
+	if not isinstance(handler, PythonHTTP):
+		vt = handler.make_viewtemplate(
+			la.DataSourceConfig(
+				identifier="persons",
+				app=config_data.apps.persons,
+			),
+			identifier="test_user_change",
+			source=f"""
+				<?print repr(globals.user.change("est123!", None, "nix@livinglogic.de"))?>
+				<?print repr(globals.user.change("Test123!", None, "nix@livinglogic.de"))?>
+				<?print globals.user.email?>
+				<?print repr(globals.user.change("Test123!", None, "xyz"))?>
+				<?code globals.user.change("Test123!", None, "walter@livinglogic.de")?>
+			""",
+		)
 
-	output = handler.renders(person_app_id(), template=vt.identifier)
-	expected = """
-		['Das alte Passwort ist falsch!']
-		[]
-		nix@livinglogic.de
-		['\"email\" muss eine gültige E-Mail-Adresse sein.']
-	"""
-	assert lines(output) == lines(expected)
+		output = handler.renders(person_app_id(), template=vt.identifier)
+		expected = """
+			['Das alte Passwort ist falsch!']
+			[]
+			nix@livinglogic.de
+			['\"email\" muss eine gültige E-Mail-Adresse sein.']
+		"""
+		assert lines(output) == lines(expected)